Add personal keyboard feature for telegram bot
Expanded the main keyboard functionality to include a personalized keyboard depending on the user's name in the spectacle. This personal menu keyboard feature is integrated in the main keyboard and triggers only if a user's full name exists in the spectacle. In addition, a new personal keyboard file has been created providing a more personalized user interaction.