Update user-specific keyboard in main_keyboard.py
Modified the keyboard layout in the `main_keyboard` function to be more user-specific. Depending on whether the user has any spectacles saved, the keyboard layout will display an option to view their spectacles or to set spectacles. The user's spectacles' title is also now used in the label of the button for viewing their spectacles.